requirements.txt
asgiref==3.7.2
Django==4.2.6
Pillow==10.1.0
pycryptodome==3.19.0
sqlparse==0.4.4
tzdata==2023.3
